Move DN tests into platform directory
